Use base16 crate and improve HMAC verification
Having a local base16_encode function is kind of dumb as noted during review.

Regarding the HMAC changes:

Previously we would compare directly against the provided hmac string.

This is fine, and we're using a constant time comparison for good crypto hygene,
but said constant time comparison panics (via an assert) if the sizes are not
identical. To keep behavior consistent with the other sync impls, we just return
false.

This also decodes the provided key from base16 before making the comparison.
This avoids an allocation but more importantly, it avoids a bug if a client
encodes the hmac key using base16 characters of a different case than we do. (In
practice everybody seems to use lowercase, but I don't think this is documented
anywhere...).

We return false immediately in the case this happens, again mimicing desktop.