Improve the token server so that it refreshes tokens and enforces backoff.
It uses a RefCell<> to manage the state (so that we don't need to declare
the functions which use the context as mutable) and uses an Enum and simple
state machine to try and get into a state where a token is available.

Needs tests, but I'm not sure how to add them (and it didn't come with tests
in the first place which I could build from)

Needs better support for errors - if there's an error fetching a token, the
caller gets a bogus error rather than the correct one.

It introduces a new StorageResetError, which could also be used in the future
if we detect a change to the syncID, and changes BackoffError to take a
SystemTime instead of an f64.