First cut at porting the Sync state machine from iOS.
This seems like a clearer model for centralizing the `remote_setup`
logic, instead of relying on the combination of observers and
error handlers that Desktop uses to figure out what to do on the
first sync, and recover from "something's wrong, check if we were
node-reassigned." This still needs to be integrated into the auth
flow. We should also figure out which object or method drives
the state machine.