Make token.rs testable and add tests (#57)
* Introduce a trait to make the TokenServer code testable, and add more tests.

I'm not at all sure if this is a sensible way to make things testable - it
seems a little awkward but gets the job done. There are some extra tests
which could be written.

* Rename TokenserverClient to TokenProvider

The public interface is now called a "TokenProvider" - it doesn't actually
provide the token itself, just services based on the token - but the name
TokenServiceProvider or similar seems unwieldy and unnecessarily long.