Build FxA Rust client in pull requests.
There are a number of changes here.  The most obvious is that we run a
decision task and build a taskgraph for PRs now.  But internally, many
things are ready to generalize to non-FxA Rust Client projects.  The
NDK and Rust toolchains are built into a Docker image (with a very
verbose name!) that is shared with Mentat.  It's nice!  The standalone
toolchains are installed in that image.  The image sets:

ANDROID_NDK_HOME=/build/android-ndk
ANDROID_NDK_API_VERSION=26
ANDROID_NDK_TOOLCHAIN_DIR=/build/android-ndk-toolchain

In $ANDROID_NDK_TOOLCHAIN_DIR the toolchains are named like
$ARCH-$ANDROID_NDK_API_VERSION; this allows to cache multiple versions
of the toolchain (if needed).  So we have, say

/build/android-ndk-toolchain/arm-26
/build/android-ndk-toolchain/arm64-26
/build/android-ndk-toolchain/x86-26

This configuration is what the `rust-android-gradle` plugin expects.
That plugin will be used for the Android Logins API in this repository.

The Dockerfile is in the Mentat repository right now, but I expect it
to migrate to application-services (or even android-components)
eventually.