[automation] Use new `libname` rust-android-gradle setting to set ELF SONAME. (Fixes #197)
This is a very different approach to producing Android-friendly
libraries than we followed in
https://github.com/mozilla/application-services/pull/175.  This
commits to always using Gradle to produce Android binaries, rather
than allowing to use Cargo (perhaps without Gradle).  It's a trade
off, and this seems better than requiring local developers to run
`patchelf` manually to test on Android 6.0.