Introduce the concept of a "stateful sync", which consumes state for a store managed externally.
Note that I'm not particularly happy with the name - I originally had it as a
"driver" with the intention of taking a vector of stores, but I'm note sure we
actually need that. So completely open to better names!

Also considering moving GlobalStateProvider to a Cell<Option<GlobalState>>
which could remove the GlobalStateProvider entirely - I figured I'd get it
up for discussion before I rearrange it yet again.