Consume pre-built OpenSSL and SQLcipher libs from Taskcluster builds.
For folks wanting to build locally, set `ext.libsGitSha = null`.  We
could grow ways to set this in `local.properties` and in any of the
other ways Gradle allows to configure the build, but it doesn't seem
important right now.  I've tried to make it possible for folks to
continue to build on Windows while we figure out cross-compiling
native code for unit tests on Windows.

This is not the most sophisticated way of achieving this, but it's
simple and pretty easy to trouble-shoot.