Produce libs/ and build `forUnitTests` artifacts for win32-x86-64 in automation.
This makes the `forUnitTests` artifacts include a .dll for
Windows (JNA's win32-x86-64) to go with the .so for Linux (JNA's
linux-x86-64) and macOS (JNA's darwin).

This uses the MingW 64-bit toolchain (i.e., the Rust target
x86_64-pc-windows-gnu) for libs/ and therefore requires it for
producing the DLL for unit testing against as well.  That's
unfortunate because local developers are probably happier building
against MSVC; but it's hard to ship libs/ for _static_ use against
MSVC out of our (Linux Docker-based) automation.

After installing MingW 64 on my Windows laptop, this works fine locally.