Part 2: Publish `-withoutLib` variants and aggregate them into composites.
This adds an Android build type that does not include the Rust native
libraries and publishes it.  The composite projects then arrange to
depend on these `-withoutLib` variants, there-by capturing the
Java/Kotlin code (and any future Android resources or assets) and any
transitive dependencies, but substituting their own libraries.

There's no way to "turn off" the parts of the `rust-android-gradle`
plugin that integrate the Rust libraries into the Android build, so we
do some manual `jniLibs` rewriting to achieve the same end.  (Features
like this in the plugin are tracked in
https://github.com/ncalexan/rust-android-gradle/issues/23).