Bump to Android-Gradle plugin 3.2.1. Fixes #484.
This is necessary to make Development with the Reference
Browser (https://mozilla.github.io/application-services/docs/applications/working-with-reference-browser.html)
smooth.

This was hard because the new kapt feature in 3.2.0+ interacts badly
with configuration resolution, and we were triggering resolution
earlier than the plugin could accommodate.  This is a generally
challenging thing to manage in Gradle, with confusing side effects;
this is one such confusing side effect!  See the in-code comments for
the details, which can't be left to a commit message.