Add a bunch of features to the integration tests.
Mainly this makes more resiliant in CI. They'll retry 5 times, with
pauses between each try. It also makes it easier to debug the helper
script (pass --helper-debug instead of modifying it's source), and
adds a few things that have been helpful in the past that I did by
manually modifying the source of the tests, e.g. keeping the account
around after the tests finish, pointing the account at a different Fxa
url, and using a specific username to run the tests.