Refactor the global state machine on the path to a SyncManager.
There are *a lot* of changes here. I think they are heading in the right direction,
but YMMV :) Note that this is still a draft, but I'd like comments on the general
approach, and any early nits to get our of the way.

The high-level changes are:

* The `GlobalState` is never persisted, and doesn't attempt to hold any
  per-collection state - it's truly *global* state. For example, it doesn't
  carry enough information to know if a store should be reset.

* There's a new `CollectionState` which is used to sync collections (eg,
  `changeset.rs` now knows only about the `CollectionState` and not the `GlobalState`).
  You obtain a `CollectionState` from a `GlobalState` and a concrete `Store`.
  The store is now expected to persist the global and collection `SyncID` it last used.
  This, in conjunction with the `GlobalState` is used to determine when an
  engine needs a reset.

* There's even migration code so that code can take the old, persisted
  `GlobalState` and extract the information it needs so engines should not need
  to be reset as part of the migration.

More details:

The `GlobalState` machine has been refactored significantly:

* There are no `Option<>` fields in the `GlobalState`, making it easier to
  rationalize about what state we are in (ie, it's impossible to have a
  "partial state")

* We don't try to recover from "strange" states - at the first whiff of things
  changing under us (eg, meta/global, keys), we restart the state machine and
  refetch *both* keys and meta/global.

* There's still a TODO for managing declined states - see the comments around
  `ApplicationState` in `state.rs` for more - I'd appreciate some thoughts
  around this because the path forward isn't particularly clear.

client.rs has been refactored too - all "GET" responses return a new Enum which
better represents the HTTP states we care about. It's now difficult to confuse
a (say) 404 for some other problem (and confusing them could be bad, and
possibly catastrophic)

The `Store`'s `reset()` function now takes the new sync IDs because we know
what the new ones are when it is called. However, there are one or 2 places
where we do not know what they are (eg, logins exposes `reset` directly) - in
that case we just use empty strings, which works fine.

`sync.rs` and `sync_multiple.rs` have been changed to better support the above,
although I expect the signatures of these will still need to change,
particularly around the `ApplicationState`.