fix(build): Set ANDROID_NDK_HOME so that gradle can strip native libs.
This is required in order for stripping of debug symbols to work
correctly, since otherwise the gradle plugin doesn't know where
to find the appropriate `strip` command.

I've also added a very coarse test to ensure that the size of the
resulting release artifacts is sensible, which should help us check
whether this works in CI and guard against future regressions.