Exclude bookmarks without URLs from `fetch_new_remote_contents`.
When we download bookmarks with missing or invalid URLs, we insert them
into `moz_bookmarks_synced` with a null `urlId`, and flag them with
`SyncedBookmarkValidity::Replace`. However, `fetch_new_remote_contents`
didn't take this into account, and would try to fetch _all_ items,
assuming that bookmarks and queries had URLs, causing the rusqlite
error in #1325. The fix is to treat the URL as optional, and ignore
bookmarks and queries that don't have one.

Closes #1325.