Implement a clients engine in the sync manager.
The clients engine handles commands for sync engines, like wiping local
data and resetting Sync metadata. These are sent in a client record's
`commands` list, and acked by removing the command from the list.

Our implementation is based on Desktop's, but works a little
differently:

* Instead of holding commands in memory, ours pulls outgoing commands
  directly from logins and Places. This follows the approach of
  bookmark restore on Desktop, which writes a flag into Places to send
  a `wipeEngine` command on the next sync.
* Our engine relies on the embedding app to supply it with info about
  the current client.

Thanks to these two changes, our engine doesn't need to store any state
itself. This means it can be used as a short-lived object that's dropped
after it syncs, like our other syncable stores, and not a long-lived
singleton as on Desktop.