Exclude hidden URLs from syncing and views.
Hidden URLs are useful for storing redirect chains, and may be useful
for manual searching or awesomebar matches. However, the history
record format doesn't currently include the referrer or hidden state,
so we exclude them entirely to match Desktop, and to avoid cluttering
other devices with redirect sources.

We also filter them out in visit queries, so that Fenix doesn't have
to make changes for mozilla-mobile/fenix#3526. In a future release,
we can either return them and have Fenix filter them out in some
views, or add a flag to include or exclude them.

See #1492.