Add explicit evaluation-order dependenies on the full-megazord.
Our gradle configuration happens to evaluate the full-megazord
project before evaluating any of the projects that depend on it,
but AFAICT that's entirely by accident - if you try to add a new
project whose name sorts earlier than "full-megazord" then gradle
will evaluate that project first, and complain when that project
tries to depend on tasks from the full-megazord build.

This commit adds an explicit dependency relationship for all the
projects that depend on full-megazord, so that when our future
selves go to copy-paste an existing `build.gradle` into a new
project named "aardvark" or whatever, we don't get a bunch of
confusing gradle errors.