Add a `Store::prepare_for_sync` method.
This method takes a clients engine, and allows the store to do whatever
it needs to prepare for syncing. The tabs store will use this to store
the current list of clients, which it uses to look up device names and
types.

Eventually, we can refactor the clients engine to also expose commands
this way, so that the store can apply any commands like resets and
wipes directly, without going through the `CommandProcessor`.

This method won't be called in the one-off sync case, which doesn't
sync the clients engine. In practice, this means that syncing tabs
without going through the sync manager will show the default device
type, and maybe stale device names.