Only attempt to sync engines if they have been configured.
Previously, if the app asked us to sync all engines then we would
attempt to sync all the engines that we know about, causing a
panic if one of those engines had not been configured.

Now, we only ever attempt to sync an engine if it has been
configured by the app.