Allow loading FxA Config objects that do not have `introspection_endpoint`.
In #2321 we added a new `introspection_endpoint` field to the FxA
RemoteConfig object, but it was added in a way that is not compatible
with previously-serialized objects. If we try to load, say, an existing
FxA state that Fenix has saved to disk, it will throw an error about the
missing `introspection_endpoint` field.

This commit makes the `introspection_endpoint` an `Option<String>` so
that we can safely deserialize state where it's missing. It also fixes
the existing tests that *should* have caught this, but didn't.

Fixes #2464.