Cache the tokenserver token and info/configuration unless there was an auth error.
Fixes #2616.

Due to a silly error we previously only cached them when a sync failed!
Fixing that is trivial but there were some TODO comments about not
dropping the cached values due to simple network or store errors
(eg, if, say, bookmarks got itself into a bad state, that shouldn't mean
these values are never cached for that user) - so now they are persisted
between syncs unless there was an auth error.

This fix also changes places-utils so it can perform multiple syncs,
meaning it's easier to see this caching behavior when running that
utility.