Rename `delete_place_by_guid` to `delete_visits_for`.
This commit changes the behavior of `delete_place_by_guid` to only
delete the page if it doesn't have any foreign key references, like
bookmarks, tags, or keywords. If it does, we can't delete the page
outright, as that would cause a constraint violation at best, and
data loss at worst. Instead, we delete all visits for the page,
but leave its row in `moz_places`, inserting tombstones if the page
is synced.

This commit also renames `delete_place_by_guid` to clarify that it
might not actually delete the page.

Closes #2520.