Reupload bookmarks with the same URL and different tags.
Tags have the same problem as keywords: Places associates them with
URLs, but Sync associates them with bookmarks. This means multiple
bookmarks with the same URL should have the same tags, but, as with
keywords, this isn't always the case. To fix this, we introduce a
query that checks if the tags for each item match the tags for its
URL. If not, we flag the item for reupload.