Fail the sync if a bookmark has an unsupported type.
We can't support bookmarks with types that we don't know about.
Ignoring them is more likely to mask client bugs, and cause problems
with the tree structure, than anything else.

Let's make this a hard error for now, and monitor telemetry. If these
occur frequently, we can explore storing placeholders for them with
`Replace` validity.