Restrict deleting from `moz_places` if there are keywords.
If a Place has keywords, we should delete them from `moz_keywords`
first, before deleting the Place from `moz_places`. This is to catch
coding errors where we accidentally ignore the `foreign_count` and
delete the Place, losing its keywords in the process.

Since we never shipped the schema in #2501 in a release, this doesn't
need a migration.