2334 - Add ttl to get_access_token.
Fixes #2334.

Adds an optional `ttl` parameter to kotlin and swift interfaces, to limit the time that tokens are cached.

The FFI uses a `zero` as a default parameter, but this is not exposed to either side of the boundary.

This is not expected to be a breaking change for the iOS and Android.