Document our decision to encourage, but not enforce, signed commits.
In the Kitchen Sync meeting earlier today, we discussed how the current
restriction of only landing signed commits on master has proved to be
an ongoing pain point without any tangible benefits. The GitHub org
security checklist says that requiring signed commits is "strongly
encouraged" and that if we don't do it, we should document the reasons why.

So, I'm documenting the reasons why and I will disable the config in
github once this documentation change lands on master.