Republish local packages if `local.properties` has changed.
The autoPublish workflow tries to avoid publishing when nothing
of consequence has changed in the repo, so it currently ignores
files that are listed in `.gitigore`. However, one of those files
(`local.properties`) can affect the result of the build.

This updates the autoPublish script to correctly re-publish if
said file has changed since the last publish.