Simplify `StorageDb` to hold a bare `Connection`.
The `webext_storage` FFI layer will hold the store in a
`ConcurrentHandleMap`, which wraps its values in a mutex, while Desktop
holds the store in a `OnceCell<Mutex<T>>`, which likewise protects the
store from concurrent access. Because of this, and the fact that the
store owns its database connection, `StorageDb` can hold an owned
reference to the `rusqlite::Connection`, without an extra layer of
`Arc<Mutex<T>>` wrapping.

Using transactions in Rusqlite was harder before, because
`Connection::transaction` required a mutable reference. Fortunately,
the latest version of Rusqlite adds `unchecked_transaction` for cases
like this, where having a `&mut Connection` is inconvenient (or
impossible).

This commit also adds some rustdoc comments to `Store` and
`StorageDb`.