Put `backtrace` behind a Cargo feature.
`backtrace` brings lots of dependencies with it, some of which
(like `gimli`) we can't vendor in Firefox Desktop because they're too
big. We could work around that, but we never log backtraces on Desktop,
anyway, and pulling in the library adds bloat. So let's put it behind
a Cargo feature flag, with a compatibility shim.