Don't reset the local ID in `TabsStore::reset()`.
If we're syncing with the sync manager, the ID will get reset when it
calls `prepare_for_sync`. If we're syncing without, `TabsEngine::sync`
will set the local ID each time it's called, before calling
`sync_multiple`.

Resetting the local ID inside of `reset` when syncing with
`TabsEngine::sync` will clear the local ID that it set, and won't get
a new ID from `prepare_for_sync` (since that never gets called in a
standalone sync), meaning we'll try to upload a record for our tabs
without an ID. This causes the server to return the 500 that we saw
in integration tests.