Integration test for sync15 component
The beginning of an integration test for the sync15 component, this
test currently inspects the behavior of sync_multiple(), storing
records and serializing them (making sure that they're parsed
correctly) and passing sync ids and metadata around (asserting that the
state machine works correctly).

There are many more aspects of sync_multiple to test, but we think that
having the scaffolding will make them easier to test. The first thing
would be to play with different arguments. Given the many moving parts
of the function, we think that doing so in an integration test is a
better start than unit tests.

Addresses #1266

Co-Authored-By: DawsonLeeRhodesCollege <dawsonleerhodescollege@users.noreply.github.com>
Co-Authored-By: Kristopher Jackson <kristopherrjackson@gmail.com>
Co-Authored-By: kwabam <kwabam@users.noreply.github.com>
Co-Authored-By: madigriffin <madigriffin@users.noreply.github.com>