Ensure setExceptionMessage gets called with a non-null value
If `setExceptionMessage` gets called with a `null` value, Prost on the
Rust side will de-serialize `exception_message` to `None`.
This is technically correct, but since there's no
`is_exception_message_defined` field that can tell this field is set,
we are relying on `exception_message` being `None` to tell us the request
failed.
This created a bug where an an invalid response would be wrongly
deserialized as a valid one.