WIP for collecting send-tab telemetry.
This is a WIP just to get feedback on the shape of this.

The main thing I'm trying to avoid here is to have the telemetry "pollute" the
public API - the public API should not be aware where and when we send
telemetry.

In this case, the "public API" in question is send-tab, where we can probably
suck up a public API change, but we shouldn't!

The whole "what telemetry API should we use/when can we use glean everywhere"
question is a good one, but one we aren't going resolve here. An ideal world
probably has our rust components directly using glean - in which case the
telemetry data also doesn't pollute the other public APIs - so let's try and
get a little closer to that world.

So the summuary of how this patch hangs together is:

* There's a new `FxaTelemetry` struct which is a grab-bag of all the telemetry
  we collect (which isn't much - even desktop today doesn't gather much). This
  struct is stored in a `RefCell<>` inside `FirefoxAccount`.

* There's a new `fxa_gather_telemetry()` function in the FFI. The idea is that
  android-components *telemetry* code will call this, and translate what it
  gets back into the glean calls it needs to make (ie, android-components will
  also add a couple of new events to `sync-telemetry/metrics.yaml` - note that
  the *send-tab* code doesn't really get involved here.

* Except that the *send-tab* code will need to tell the telemetry code that
  there's probably telemetry to gather. So while it's not magic or fully
  automatic, most of the responsibilities are in the right place and there's
  only a few leaky abstractions.

Assuming the "big picture" of this is OK, some "small picture" issues:

* I'm not sure if it's OK that `fxa_gather_telemetry()` returns a JSON string.
  The idea is that it should ideally be dynamic - stuff it doesn't understand
  should be ignored. Eg, you can imagine a world where some things are only
  understood/recorded by, say, iOS. Adding new telemetry shouldn't be a breaking
  change to platforms that don't want to explicitly record it.

* No tests!

* Um - I'm sure I had more than that :) I'm sure you'll find some in the patch.