Delete everything from `master` branch except an explanatory note.
We've moved to `main` as our default branch, but some folks may have
old checkouts that reference the `master` branch or may accidentally
use it from muscle-memory. This commit tries to help them out by
deleting all the useful stuff from the `master` branch and replacing
it with a note telling them to use `main`.