Port the sql function note_sync_change() from desktop.
This allows us to avoid applying the result of a bookmark sync merge
if the local tree has changed since we fetched it.

It's named `note_bookmarks_sync_change()` here to better
reflect the fact that history and bookmarks are much tighter
here than on desktop and it applies only to bookmarks, but
otherwise it is as close to the desktop implementation as
possible.