Don't implicitly update experiments on first use. (#67)
Previously, we would implicity call `update_experiments` when trying
to list the experiments in an empty database. We would not update
if the database contained at least one experiment, no matter how stale.

In practice this seems to be a bit of a footgun, because it looks
to consumers like we'll automatically update definitions when
initialising the client, but this only works once.

Instead, let's require consumers to make an explicit call to
`update_experiments` whenever they want to update the experiments.