Rename 'engine' and 'store' to conform with the sync15 readme.
In many cases, this was literally changing 'engine' to 'store' and
'store' to 'engine', but now we have mostly consistent naming
throughout the repo, and consistency with desktop.

The sync trait is now called `SyncEngine` just to drive home the
fact that 'engine' relates to syncing.

This was largely a mechanical change. No changes were made to
the FFI functions, so this is not a breaking change.