Add `reset_telemetry_identifiers` method.
When the user disables telemetry in the consuming application, we need
to reset any unique identifiers used internally by the SDK. This guards
against accidentally correlating metrics submitted before the opt-out
with metrics that the user may opt-in to submitting again in future.

Consumers are expected to call `reset_telemetry_identifiers` when the user
disables telemetry in order to trigger resetting of the identifiers.
They may also call it when the user re-enables telemetry in order to
provide updated randomization unit values.