Refactor DB to remove mutex, use more explicit transactions.
This commit started out with the observation that the Rkv `Database`
class is `Sync` and `Send`, so we don't need to wrap it in a separate
mutex in order to coordinate use from multiple threads. One less mutex
should be a nice cleanup, right?

Turns out we were depending on that mutex for guarding against stale
updates in the database caching layer, so I've refactored that to make
it depend instead on the caller passing in an explicit write transaction.

In the process of doing *that*, I discovered that the way we manage
database transactions seems pretty haphazard. Some functions take just
a `&Database` and manage their own transaction, while their peers take
both `&Database` and `&mut Writer`. We even had a read-only function
that was taking a `&mut Writer` purely so that it could sometimes be
called from within an existing transaction (but sometimes not).

This way lies madness.

So, the bulk of this commit is actually me trying to impose a little
order on how we manage database transactions, based on the following
principles:

* Manage transactions at the outermost layer only. All helper functions
  should take an existing transaction handle, and only the top-level
  client that owns the database should be managing the lifecycle of
  transactions.
* Don't take a write transaction if you don't need one. This has
  required a bit of trait fiddliness but I think it's worthwhile
  to accurately communicate intent (and allow the future possibility
  of more read parallelism).