autofill: Use a custom struct for serializing and deserializing sync payload
Because Address and CreditCard records on the server have an "odd" shape,
reusing the InternalAddress and InternalCreditCard structs for Sync was
becoming more trouble than it's worth - custom structs are a little
more code, but do make for a cleaner separation of concerns (eg, the
'version' field from the 'Internal' structs has been removed)

This meant that the serialization support from the InternalStructs
could be removed - it might be more confusing to have these
structs being serializable from JSON when there's no actual
use case for it - people might assume it's for Sync.

Except the autofill sample was the one use-case :) I toyed with
the idea of having that also use the sync15 payload, but that
seems wrong and would mean making more things pub than I'd like,
so I bit the bullet and changed that to "prompt" for values just
like the logins sample does.

With this patch, the sample syncs and correctly applies the records!