Remove unused gradle build config from this repo. (#106)
All consumers of the Nimbus Rust SDK are expected to get it via the
Application Services framework. In fact, if they try to use a version
of Nimbus that wasn't build as part of the Application Services
"megazord" then it will not actually work, because the networking
layer will not have been properly initialized.

To ease maintenance and avoid future confusion, let's delete the
standalone android publishing config from the nimbus-sdk repo,
since we never expect anyone to consume the artifacts it produces
in practice.

Since we're not producing built artifacts from this repo, there is
no point in checking in the `Cargo.lock` file for the Rust crate,
so that has also been removed.