autofill: add credit-card encryption.
Changes the credit-card schema to have `cc_number_enc`, which is encrypted
using jwcrypto, and `cc_number_last_4` which is not encrypted.

This requires the embedding application to perform all the crypto, so
there are new public functions for encrypting, decrypting and creating
new keys. Further, syncing also requires the key, so this is able to be
passed in when doing a sync.