Update to Glean v36.0.0
This brings a breaking change (but it's a bugfix) for iOS consumers:

Event timestamps are now correctly recorded in milliseconds.
Since the first release event timestamps were erroneously recorded with nanosecond precision.
This is now fixed and event timestamps are in milliseconds. This is equivalent to how it works in all other language bindings.