Automagically manage rust version and targets via toolchain config.
We originally landed this change in #3936 but had to revert it due
to release build issues. As far as I can tell, those were due to
mismatches between our use of a rust-toolchain file for repos-specific
config, and our CI scripts that try to configure the default rustup
toolchain.

This commit moves as much rustup config as possible into the toolchain
file, where we can maintain it in a single consistent place.