Add a walkthrough for adding a new component from scratch.
There are a lot of non-obvious things you need to do when adding
a new component to this repo. I've tried to capture them all
in one reference guide, including screenshots for the part
where you have to mess around in the XCode GUI.

There are probably parts of this that could be simplified,
but let's start by writing the whole process down.