Add a full "how to add a component" howto (and various other doc updates) [ci skip] (#3987)
* Add a walkthrough for adding a new component from scratch.

There are a lot of non-obvious things you need to do when adding
a new component to this repo. I've tried to capture them all
in one reference guide, including screenshots for the part
where you have to mess around in the XCode GUI.

There are probably parts of this that could be simplified,
but let's start by writing the whole process down.

* Delete the "using protobuf-encoded data over Rust FFI" howto.

We have some existing components that use this approach, but any
new components should use UniFFI's builtin "record" types for
passing data rather than using protobuf. Thus, we don't expect
anyone to actually need this howto in future.

* Morph the "exposing rust code on android" howto into a simpler FAQ.

The existing doc was a hybrid howto/FAQ already, and we intend for
the "howto" part to be entirely replaced by "use UniFFI". So, let's
salvage content that's still useful to us and put it in a deliberate
FAQ.

* Remove the "consuming rust components on android" howto.

It's basically content-free, and the one link it has there about
megazords has been dead for quite some time. This doc is obviously
not providing any value in practice.

* Remove the "exposing rust components to swift" howto.

It's pretty meagre, and it was pretty much all about the megazord config changes,
which are now covered by the general "adding a new component" guide.