Adding test_mod test
Adding a test case for scrubbed data, although doesn't seem like these
tests are actually being run.  In fact, I get errors when I try to
import the tests module.  Still, I'm pretty sure this test is correct,
and hopefully it will be useful if/when we get the code here running
again.