Add documentation for publishing pre-built NSS artifacts.
We currently rely on pre-built binary artifacts for running NSS
tests on Mac and Windows machines. The procedure for generating
them is outlined in #962, but if we're being honest with ourselves,
it is a little bit of a "How to Draw an Owl" situation...

I took some time today to try to walk through the procedure, fill
in the missing steps, and document it in more detail. This commit
updates the "upgrading NSS" howto document with the result.

We should continue to work on not having so many steps involved
in this process, but having the current steps spelled out is
a good first step!