CircleCI must install Python in order to build system libs.
The docker image change in #4110 seems to mean that `python` is no
longer available in the Docker image by default, which means that
the build scripts in `./libs` no longer run successfully on CircleCI.

We did not notice this change because CircleCI usually restores
these from a cache. But if we do something that invalidates that
cache (like modifying the files in `./libs`) then CircleCI starts
failing.

This commit fixes the issue by explicily installing `python` as
part of the dependencies for the system lib builds. It also includes
a cosmetic change to `./libs/README.md` to clear the CircleCI cache
and test that the fix actually works.