Use a shared helper for gradle uniffi-bindgen tasks.
We currently have four copies of the logic for calling UniFFI
from gradle, several with their own little tweaks. Let's take
the best lessons from each and consolidate them into a single
shared helper.

This version is, I think, an improvement over all the existing
ones:

* Like the nimbus version, it registers the task as generating
  java source files, which should help things integrate cleanly
  into the broader build.
* Like the fxa-client version, it declares explicit inputs so
  that we can avoid re-running the task when nothing has changed.
* New in this version, it declares an input dependency on anything
  that would affect the version of `uniffi-bindgen` that is in use,
  since that too will change the generated outputs.

That last change should really help with smooth updates to the
version of UniFFI in use. Previously, if you changed the UniFFI
version it would not trigger a cleanup of old generated bindings,
and you'd get `java.lang.UnsatisfiedLinkError` errors like in #4054.
(This is UniFFI's last-line-of-defence measure to prevent you from
using bindings generated with one version of UniFFI on a library
compiled with another, which is very useful, but an extremely
confusing failure mode).

With this change, an updated version of UniFFI will invalidate
the gradle task that generates the bindings, causing them to
be regenerated with the new version.