Made MigrationLogic a trait instead of a struct
rfk suggested this to usage feel a bit more declarative and I like it
too.

It also allows associating some state with the upgrade functions.  This
is great for the tests, although I'm not sure if we'll need that for
real examples.

Moved error_handling from a field on the MigrationLogic struct to a
parameter in open_database().  Mostly this was because it made things
easier to test, but I also think it makes sense from a design point of
view.  The error handling is orthogonal to the migration logic.